# docs/swagger.yaml
definitions:
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /reading-lists/{list_id}:
    get:
      description: Retrieve a single reading list. Public lists are accessible without
        auth; private lists require the owner or a shared user.
      parameters:
      - description: Reading list ID
        in: path
        name: list_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reading list data
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: List not found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get reading list by ID
      tags:
      - reading-lists
  /reading-lists/{list_id}/manga:
    post:
      consumes:
      - application/json
      description: Append a manga to one of your own reading lists
      parameters:
      - description: Reading list ID
        in: path
        name: list_id
        required: true
        type: string
      - description: 'Manga to add — {manga_id: string}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Manga added
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Manga already in list
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Not your list
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Add manga to reading list
      tags:
      - reading-lists
  /reading-lists/{list_id}/manga/{manga_id}:
    delete:
      description: Remove a specific manga from one of your reading lists
      parameters:
      - description: Reading list ID
        in: path
        name: list_id
        required: true
        type: string
      - description: Manga ID to remove
        in: path
        name: manga_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Manga removed
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Not your list
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Manga not in list
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Remove manga from reading list
      tags:
      - reading-lists
  /reading-lists/{list_id}/subscribe:
    delete:
      description: Stop following a reading list you previously subscribed to
      parameters:
      - description: Reading list ID
        in: path
        name: list_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Unsubscribed successfully
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Not subscribed
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Unsubscribe from a reading list
      tags:
      - reading-lists
    post:
      description: Follow a public reading list owned by another user
      parameters:
      - description: Reading list ID
        in: path
        name: list_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subscribed successfully
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Already subscribed or own list
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: List not found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Subscribe to a reading list
      tags:
      - reading-lists
  /reading-lists/create:
    post:
      consumes:
      - application/json
      description: Create a new reading list with a title, description, and initial
        manga IDs
      parameters:
      - description: 'Reading list — {title, description, manga_list: [], is_public:
          bool}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Reading list created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Missing title or manga_list
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create reading list
      tags:
      - reading-lists
  /reading-lists/mine:
    get:
      description: Return all reading lists owned by the authenticated user
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Results per page (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: My reading lists
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get my reading lists
      tags:
      - reading-lists
  /reading-lists/public:
    get:
      description: List all publicly shared reading lists with pagination (no auth
        required)
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Results per page (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Public reading lists
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Browse public reading lists
      tags:
      - reading-lists
  /reading-lists/subscribed:
    get:
      description: Return all public reading lists that the authenticated user has
        subscribed to
      produces:
      - application/json
      responses:
        "200":
          description: Subscribed lists
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get subscribed reading lists
      tags:
      - reading-lists
...
swagger: "2.0"